# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "chimera-bench server API",
    "version": "0.0.1"
  },
  "components": {
    "schemas": {
...
      "WorkloadsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "type": "object",
            "properties": {
              "workloads": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "workloadId": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "displayName": {
                      "type": "string",
                      "minLength": 1
                    },
                    "version": {
                      "type": "string",
                      "minLength": 1
                    },
                    "promptCount": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "source": {
                      "type": "string",
                      "enum": [
                        "built-in",
                        "filesystem"
                      ]
                    }
                  },
                  "required": [
                    "workloadId",
                    "displayName",
                    "version",
                    "promptCount",
                    "source"
                  ]
                }
              }
            },
            "required": [
              "workloads"
            ]
          },
          "meta": {
            "type": "object",
            "properties": {
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "requestId"
            ]
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      },
      "WorkloadDetailResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "type": "object",
            "properties": {
              "workloadId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 128
              },
              "displayName": {
                "type": "string",
                "minLength": 1
              },
              "version": {
                "type": "string",
                "minLength": 1
              },
              "promptCount": {
                "type": "integer",
                "minimum": 0
              },
              "source": {
                "type": "string",
                "enum": [
                  "built-in",
                  "filesystem"
                ]
              },
              "promptIds": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                }
              },
              "prompts": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "promptId": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "caseId": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "messages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "role": {
                            "type": "string",
                            "enum": [
                              "system",
                              "user",
                              "assistant"
                            ]
                          },
                          "content": {
                            "type": "string",
                            "minLength": 1
                          }
                        },
                        "required": [
                          "role",
                          "content"
                        ]
                      },
                      "minItems": 1
                    },
                    "contextFiles": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "notes": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "promptId",
                    "caseId",
                    "messages"
                  ]
                }
              }
            },
            "required": [
              "workloadId",
              "displayName",
              "version",
              "promptCount",
              "source",
              "promptIds"
            ]
          },
          "meta": {
            "type": "object",
            "properties": {
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "requestId"
            ]
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      },
      "WorkloadsReloadResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "type": "object",
            "properties": {
              "discoveredPacks": {
                "type": "integer",
                "minimum": 0
              },
              "skippedInvalidPacks": {
                "type": "integer",
                "minimum": 0
              },
              "duplicateIdSkips": {
                "type": "integer",
                "minimum": 0
              }
            },
            "required": [
              "discoveredPacks",
              "skippedInvalidPacks",
              "duplicateIdSkips"
            ]
          },
          "meta": {
            "type": "object",
            "properties": {
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "requestId"
            ]
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              }
            },
            "required": [
              "code",
              "message"
            ]
          },
          "meta": {
            "type": "object",
            "properties": {
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "requestId"
            ]
          }
        },
        "required": [
          "success",
          "error",
          "meta"
        ]
      }
    },
...
  },
  "paths": {
...
    "/workloads": {
      "get": {
        "summary": "List registered workloads",
        "tags": [
          "workloads"
        ],
        "responses": {
          "200": {
            "description": "Registered built-in and filesystem workloads",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkloadsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workloads/{workloadId}": {
      "get": {
        "summary": "Read workload metadata or prompt details",
        "tags": [
          "workloads"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "required": true,
            "name": "workloadId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            },
            "required": false,
            "name": "includePrompts",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Workload details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkloadDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params or query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workload not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Response payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workloads/reload": {
      "post": {
        "summary": "Reload filesystem workloads",
        "tags": [
          "workloads"
        ],
        "responses": {
          "200": {
            "description": "Reload complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkloadsReloadResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Reload cooldown active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
...
}